livedataservice/requirements.txt
astroid==2.11.5
async-timeout==4.0.2
autopep8==1.6.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
cleo==0.8.1
clikit==0.6.2
crashtest==0.3.1
Deprecated==1.2.13
dill==0.3.4
distlib==0.3.4
filelock==3.7.0
grpc-tools==0.0.1
grpcio==1.46.3
grpcio-tools==1.46.3
html5lib==1.1
idna==3.3
importlib-metadata==4.11.4
isort==5.10.1
keyring==23.5.1
lazy-object-proxy==1.7.1
lockfile==0.12.2
mccabe==0.7.0
msgpack==1.0.3
mypy==0.950
mypy-extensions==0.4.3
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
pkginfo==1.8.2
platformdirs==2.5.2
poetry-core==1.0.8
protobuf==3.20.1
ptyprocess==0.7.0
pycodestyle==2.8.0
pylev==1.4.0
pylint==2.13.9
pyparsing==3.0.9
python-dotenv==0.20.0
redis==4.3.1
requests==2.27.1
requests-toolbelt==0.9.1
shellingham==1.4.0
six==1.16.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
types-redis==4.2.5
types-requests==2.27.27
types-urllib3==1.26.14
typing-extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
webencodings==0.5.1
wrapt==1.14.1
zipp==3.8.0
Flask==2.1.2
